stress testing along with periodic sync
